Separate range module
